+++ test/EFCore.Benchmarks/Models/AdventureWorks/EmployeeDepartmentHistory.cs
@@ -9,7 +9,7 @@ namespace Microsoft.EntityFrameworkCore.Benchmarks.Models.AdventureWorks
     {
         public int BusinessEntityID { get; set; }
         public DateTime StartDate { get; set; }
-        public short DepartmentID { get; set; }
+        public int DepartmentID { get; set; }
         public byte ShiftID { get; set; }
         public DateTime? EndDate { get; set; }
         public DateTime ModifiedDate { get; set; }
